Modified the size attribute so that it allows decimal numbers Fixed a bug in the web results page => the total count includes only checked attributes In the web match page, allow the display of ascii coded text

+++ MycoKeyCreator.Library/DBObject/SpeciesAttributeSize.cs
@@ -10,6 +10,6 @@ namespace MycoKeyCreator.Library.DBObject
         public Int64 key_id { get; set; }
         public Int64 species_id { get; set; }
         public Int64 attribute_id { get; set; }
-        public Int16 value { get; set; }
+        public float value { get; set; }
     }
 }
