+++ MycoKeyCreator.WebApplication/Model/KeyMatchViewOutput.cs
@@ -15,7 +15,7 @@ namespace MycoKeyCreator.WebApplication.Model
             public bool IsSelected { get; set; }
             public Int64 AttributeId { get; set; }
             public Int16 AttributeType { get; set; }
-            public Int64 AttributeValue { get; set; }
+            public float AttributeValue { get; set; }
         }
 
         public string KeyName { get; set; }
